Move to write_locks file
