+++ code/go/0chain.net/blobbercore/writemarker/write_locks.go
@@ -1,4 +1,4 @@
-package datastore
+package writemarker
 
 import "time"
